beginning work on TVService
